+++ src/django_safemigrate/__init__.py
@@ -5,3 +5,4 @@ class Safe(Enum):
     always = "always"
     before_deploy = "before_deploy"
     after_deploy = "after_deploy"
+    unset = "unset"